% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_inline.R
\name{as_inline}
\alias{as_inline}
\title{Convert table data to inline list}
\usage{
as_inline(data, table_names, table_value)
}
\arguments{
\item{data}{a data frame.}

\item{table_names}{column names that will be used to form groups in the table}

\item{table_value}{column name that contains table values.}
}
\value{
a list of \code{table_value} values for each permutation of \code{table_names}
}
\description{
Convert table data to inline list
}
\note{
variables in \code{table_names} that have missing values will be
have their missing values converted into an explicit category named
variable_missing, where 'variable' is the name of the variable.
}
\examples{

example_data <- data.frame(
  sex = c("female", "male"),
  height = c("158 (154 - 161)", "178 (175 - 188)")
)

as_inline(example_data, table_names = 'sex', table_value = 'height')

car_data <- mtcars
car_data$car_name <- rownames(mtcars)
as_inline(car_data, table_names = 'car_name', table_value = 'mpg')

}
