% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_spec.R
\name{default_rounder_set}
\alias{default_rounder_set}
\alias{default_rounder_reset}
\title{Set the default rounding specification}
\usage{
default_rounder_set(rspec)

default_rounder_reset()
}
\arguments{
\item{rspec}{a \code{rounding_specification} object (see \link{round_spec}).}
}
\value{
\code{default_rounder_set()} changes global options and does
not return a value.
}
\description{
If you ever get tired of having to write
\code{table_glue(..., rspec = your_rspec)} and would rather just write
\code{table_glue(...)}, you can use the \code{default_rounder_set()} function
on \code{your_rspec} to temporarily update \code{options} in your current R
session and update the default rounding specification to match
whatever your current rounding specification is.
Naturally, mistakes will be made when \code{default_rounder_set()} is used.
To reset the rounding specifications to factory defaults,
use \code{default_rounder_reset()}.
}
\details{
Setting a default rounding specification will impact
all functions in the \verb{table helpers} family. Use with caution! Also,
please make all use of \code{default_rounder_set()} \strong{explicit}
in your code so that it can be reproduced. It is highly recommended
that you \strong{do not} use \code{default_rounder_set()} in your .Rprofile.
}
\examples{

your_rspec <- round_using_decimal(round_spec(), digits = 10)
your_rspec <- format_small(your_rspec, mark = '--')

# now your_rspec is the default.
default_rounder_set(your_rspec)

# so it is applied whenever you use table_value(),
# unless you specify otherwise.
table_value(rnorm(5))

# don't forget to reset when you're done
default_rounder_reset()

}
