% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_spec.R
\name{format_decimal}
\alias{format_decimal}
\title{Format decimal symbol}
\usage{
format_decimal(rspec, mark = ".")
}
\arguments{
\item{rspec}{a \code{rounding_specification} object (see \link{round_spec}).}

\item{mark}{a character value used to separate number groups to the
left of the decimal point. See \link{prettyNum} for more details on this.
Set this input to '' to negate it's effect.}

\item{interval}{a numeric value indicating the size of number groups
for numbers left of the decimal.}
}
\value{
an object of class \code{rounding_specification}.
}
\description{
The decimal symbol can be update.
}
\examples{

small_x <- 0.1234567

rspec <- round_spec()
rspec <- round_using_decimal(rspec, digits = 7)
rspec <- format_small(rspec, mark = '*', interval = 1)

table_value(small_x, rspec)

}
