% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_spec.R
\name{format_missing}
\alias{format_missing}
\title{Format missing values}
\usage{
format_missing(rspec, replace_na_with)
}
\arguments{
\item{rspec}{a \code{rounding_specification} object (see \link{round_spec}).}

\item{replace_na_with}{a character value that replaces missing values.}
}
\value{
an object of class \code{rounding_specification}.
}
\description{
\code{format_missing()} updates a \code{rounding_specification} object so that
missing values are printed as the user specifies.
}
\examples{

rspec <- round_spec()
rspec <- format_missing(rspec, 'oh no!')
table_value(x = c(pi, NA), rspec)

}
