% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_spec.R
\name{make_default_rounder}
\alias{make_default_rounder}
\alias{reset_default_rounder}
\title{Set the default rounding specification}
\usage{
make_default_rounder(rspec)

reset_default_rounder()
}
\arguments{
\item{rspec}{a \code{rounding_specification} object (see \link{round_spec}).}
}
\value{
\code{make_default_rounder()} changes global options and does
not return a value.
}
\description{
If you ever get tired of having to write
\code{table_glue(..., rspec = your_rspec)} and would rather just write
\code{table_glue(...)}, you can use the \code{make_default_rounder()} function
on \code{your_rspec} to temporarily update \code{options} in your current R
session and update the default rounding specification to match
whatever your current rounding specification is.
Naturally, mistakes will be made when \code{make_default_rounder()} is used.
To reset the rounding specifications to factory defaults,
use \code{reset_default_rounder()}.
}
\details{
Setting a default rounding specification will impact
all functions in the \verb{table helpers} family. Use with caution! Also,
please make all use of \code{make_default_rounder()} \strong{explicit}
in your code so that it can be reproduced. It is highly recommended
that you \strong{do not} use \code{make_default_rounder()} in your .Rprofile.
}
\examples{

your_rspec <- round_using_decimal(round_spec(), digits = 10)
your_rspec <- format_small(your_rspec, mark = '--')

# now your_rspec is the default.
make_default_rounder(your_rspec)

# so it is applied whenever you use table_value(),
# unless you specify otherwise.
table_value(rnorm(5))


}
