% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.glue-package.R
\docType{package}
\name{table.glue-package}
\alias{table.glue}
\alias{table.glue-package}
\title{table.glue: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
