% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.glue-package.R
\docType{package}
\name{table.glue-package}
\alias{table.glue}
\alias{table.glue-package}
\title{table.glue: Make and Apply Customized Rounding Specifications for Tables}
\description{
Translate double and integer valued data into
    character values formatted for tabulation in manuscripts
    or other types of academic reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcjaeger/table.glue}
  \item Report bugs at \url{https://github.com/bcjaeger/table.glue/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

}
\keyword{internal}
